Reviewed by:	stas
For the benefit of Win32, use the new function t_filepath_cmp()
in Apache::TestUtil to compare two filepaths. On non-Win32 this
is the same as t_cmp(), but on Win32, this first converts both paths
to their DOS long pathname before invoking t_cmp(). This avoids
spurious failures in cases when one of the paths is represented
by its long name and the other by its short name.


git-svn-id: https://svn.apache.org/repos/asf/httpd/test/trunk/perl-framework/Apache-Test@103821 13f79535-47bb-0310-9956-ffa450edef68